backend/requirements.txt
asgiref==3.5.2
django==4.0.6
gitdb==4.0.9
gitpython==3.1.27
smmap==5.0.0
sqlparse==0.4.2
djangorestframework==3.13.1